Rename presentational CommandModal component to distinguish from web app component
