+++ packages/react-ui/src/Modals/index.ts
@@ -1,6 +1,6 @@
 import useManagedWaypoints from './MissionModalSteps/hooks/useManagedWaypoints'
 
-export * from './CommandModal'
+export * from './CommandModalView'
 export * from './ConfirmStopModal'
 export * from './NewDeploymentModal'
 export * from './LoginModal'
